Quitado el @Future de challenge y añadida validacion en los servicios, cambias estrellas de investor record de integer a double

+++ src/main/population/create-schema.sql
@@ -91,7 +91,7 @@
         `investing_statement` double precision,
         `name` varchar(255),
         `sector` varchar(255),
-        `stars` integer,
+        `stars` double precision,
         primary key (`id`)
     ) engine=InnoDB;
